requirements.txt
rumps
pyobjc
kucoin-universal-sdk
websocket-client
